+++ utility/install_os_dependencies.sh
@@ -31,8 +31,3 @@ echo "***********************************************"
 echo "Installing django-extensions dependencies"
 echo "***********************************************"
 apt-get -y install graphviz-dev
-
-echo "***********************************************"
-echo "Installing SAML dependencies"
-echo "***********************************************"
-apt-get -y install xmlsec1