+++ CHANGELOG.rst
@@ -30,6 +30,7 @@ Removed
     - Python v3.11 support (#1994)
     - django-docs dependency (#1996)
     - versioneer dependency (#1995)
+    - xmlsec1 dependency for previously removed SAML support (#1992)
 - **Projectroles**
     - Kiosk mode support (#1775)
     - ``PROJECTROLES_KIOSK_MODE`` env var (#1775)